Final Demo Submission: Virtual network Peering in Microsoft Azure
### Final Demo Submission: Virtual network Peering in Microsoft Azure #470

**Members**
Helal Uddin (mhud@kth.se)
Adam Björnberg (adambjo@kth.se)

**Screencast Link**: https://youtu.be/o7DkXFiBHjk **(Please choose Quality 1080 HD option in the youtube setting option to see it clearly)**

**Description:**
According to Microsoft, Virtual network peering enables a seamless connection between Azure virtual networks. By using both private and global virtual network peering, it is possible to route traffic with low-latency, high-bandwidth. Virtual machines can communicate with other virtual machines by using VNP. This communication process is connected with other resources such as Azure subscription, Azure Active Directory(AAD) in the Azure portal. As this communication is private, there is no necessity to implement encryption. In this demo, we will present how to create multiple virtual machines, virtual network peering in the Azure portal.


**Grading Criteria fulfilled by this demo:**

- The demonstration is clearly motivated (why it matters for Devops?) | **Yes** | No | **Remarkable**
Azure DevOps and Virtual network peering are closely connected with each other as virtual network peering provides a seamless connection between Azure virtual networks and Azure DevOps.

- The demonstration is difficult to do | **Yes** | No | **Remarkable**
The demo demonstration is difficult to do for the beginners and also in a limited time limit. To create a virtual machine and delete a resource group, it takes at least 15 to 20 minutes. But due to the global pandemic, it could be more than that. As per our demo screencast, we tried to present it less than 5 minutes which is very difficult to do.

- The demonstration is original | **Yes** | No | Remarkable
The demo is created by the members and it's subscription id and resource groups are original. 
Though, from the Microsoft documentation, a similar type of information could be found. Here is the reference: https://docs.microsoft.com/en-us/azure/virtual-network/tutorial-connect-virtual-networks-portal

- The demonstration is sublime (eg visually appealing) | **Yes** | No | **Remarkable**
This demonstration is visually appealing as Virtual networks and virtual machines are very important parts of the Azure cloud communication and most of the services are connected with it. As the whole VM and VN are presented in the modern user interface, it is remarkably appealing to the users.

- The live demonstration contains an easter egg | **Yes** | No | Remarkable
The topic is very interesting and easter egg/fun fact has been added.

- The audience can play / directly interact with the demo | **Yes** | No | **Remarkable**
The user of the VM can easily connect with the virtual network by using the ping command remotely. Because of this remote interaction of Virtual Machines, it remarkably ensures both secured and express communication between networks. 

- The demonstration triggers a conversation | **Yes** | No | Remarkable
This demonstration obviously triggers a conversation on the audience regarding how to connect different networks by using a peering mechanism and how to integrate many VM devices in the network.

- A screencast has been made (optional) | **Yes** | No | Remarkable
A screencast has been made to present the demonstration of the Azure virtual network peering.

This screencast contains subtitles to make it clear for the audience that at the beginning, the first virtual network has been created. Next, the second virtual network has been created. Following that, network peering has been created in the Virtual network 1 and added VM into it. To make it clear, subtitles are visible under the screencast and change it according to the screencast. Finally, Take home message has been given by stating that deleting a resource group can help to reduce unnecessary costs and that has to be done after competing for demo.