Add presentation proposal for week 6 (#2194)
# Assignment Proposal

## Title

How [MLFlow](https://mlflow.org/docs/latest/concepts.html) can enhance your ML operations.

## Names and KTH ID

  - Omid Hazara (hazara@kth.se)
  - Linus Östlund (linusost@kth.se)

## Deadline

- Week 6

## Category

- Presentation

## Description

We are going to present the core concepts of MLFlow and how it is used to improve ML Operations.
MLFlow is open-source platform, and a popular tool to manage the end-to-end ML life cycle. 
It is divided into components, such as tracking, models, projects, model registry and model serving. 
Our goal is to show how these components benefit ML-modelling, training and serving. 

**Relevance**

Using a platform such as MLFlow helps practitioners (data engineers, analysts, etc) to collaborate with less effort. 
Tracking experiments, storing the results in a central repository, and reproducing them is key to successful machine learning.
MLFlow also offers a way to package and deploy models, which is of equal importance.