End-to-end testing for webapps using Playwright
- Tobias Ljunggren (tljun@kth.se)
- Josef Lindkvist (joslindk@kth.se)

- Task 1, Deadline 1 for async tasks: Oct 1 2024, 17h Stockholm time

- Executable Tutorial

Our executable tutorial will include end-to-end testing for webapps utilizing Playwright (https://playwright.dev/). It will be an executable demonstration on how you can use Playwright in different ways to enhance your end-to-end tests.
We will upload this tutorial to Google Colab.

_Relevance_

This tool is an end-to-end frontend testing framework that is widely
used in the industry. We believe that this tutorial could be useful for
CS students and others who are not very versed in software testing tools
for the web.