Week 6: Demo proposal (#2393)
* Executable tutorial project proposal, Playwright testing

This is a project proposal for an executable tutorial by Josef Lindkvist
and Tobias Ljunggren.

* End-to-end testing for webapps using Playwright

- Tobias Ljunggren (tljun@kth.se)
- Josef Lindkvist (joslindk@kth.se)

- Task 1, Deadline 1 for async tasks: Oct 1 2024, 17h Stockholm time

- Executable Tutorial

Our executable tutorial will include end-to-end testing for webapps utilizing Playwright (https://playwright.dev/). It will be an executable demonstration on how you can use Playwright in different ways to enhance your end-to-end tests.
We will upload this tutorial to Google Colab.

_Relevance_

This tool is an end-to-end frontend testing framework that is widely
used in the industry. We believe that this tutorial could be useful for
CS students and others who are not very versed in software testing tools
for the web.

* [200~# Assignment Proposal

End-to-end testing for webapps using Playwright

- Tobias Ljunggren (tljun@kth.se)
- Josef Lindkvist (joslindk@kth.se)

- Week 4

- Demo

Our demo will include end-to-end testing for web apps utilizing
Playwright (https://playwright.dev/). It will demonstrate a couple of
smart ways to use Playwright to enhance your end-to-end tests. In our
demo, we will start by showing how quickly and easily you can get
started with Playwright. We will then live-write a few frontend tests in
a pre-prepared repository. By running the tests, we will show how
Playwright works and utilizes cross-browser testing. Finally, we plan to
demonstrate how you can add Playwright testing to your CI pipeline with
GitHub Actions.

Plan B:
If something doesn't work during our demo we will have screen recordings
prepared to show.

This tool is an end-to-end frontend testing framework that is widely
used in the industry. We believe that this tutorial could be useful for
CS students and others who are not very versed in software testing tools
for the web.

End-to-end testing for webapps using Playwright

- Tobias Ljunggren (tljun@kth.se)
- Josef Lindkvist (joslindk@kth.se)

- Week 4

- Demo

Our demo will include end-to-end testing for web apps utilizing Playwright (https://playwright.dev/). It will demonstrate a couple of smart ways to use Playwright to enhance your end-to-end tests. In our demo, we will start by showing how quickly and easily you can get started with Playwright. We will then live-write a few frontend tests in a pre-prepared repository. By running the tests, we will show how Playwright works and utilizes cross-browser testing. Finally, we plan to demonstrate how you can add Playwright testing to your CI pipeline with GitHub Actions.

Plan B:
If something doesn't work during our demo we will have screen recordings prepared to show.

This tool is an end-to-end frontend testing framework that is widely used in the industry. We believe that this tutorial could be useful for CS students and others who are not very versed in software testing tools for the web.

* changed name of folder

* formatting

* new proposal

* new week proposal

* new week proposal

---------

Co-authored-by: Josef Lindkvist <joslindk@kth.se>