Create a Spotify playlist with Terraform
- Josef Lindkvist (joslindk@kth.se)
- Josephine Kuo (jkuo@kth.se)

- Task 3

- Executable Tutorial

In this executable tutorial, we will showcase how to utilize Terraform. Terraform manages infrastructure on cloud computing providers by building, changing, and versioning infrastructure. In this tutorial, we will create a Spotify playlist using Terraform. The focus will be on Infrastructure as Code principles, showcasing how Terraform can manage and provision cloud infrastructure. We'll use Docker to contain the environment, and Google Colab to execute the code.

**Relevance**
This tutorial is very relevant to the topic of “infrastructure as code” and thus DevOps because it showcases how Terraform can help us define, control, and automate infrastructure. Spotify is a good example of this because its API offers a well-structured, real-world service that requires authentication, permissions, and interactions with an external platform. This mirrors the tasks that DevOps engineers have to perform where automation and provisioning must interact with third-party services.